reorganize component folder
