+++ frontend/app/(app)/notification/components/skeleton.tsx
@@ -1,0 +1,29 @@
+import { AnimatedBox } from "@/design-system/base/animated-box";
+import usePulsingAnimation from "@/hooks/component/pulse-animate";
+import { Animated, View } from "react-native";
+
+const NotificationSkeleton = () => {
+  const opacity = usePulsingAnimation();
+
+  return (
+    <AnimatedBox
+      borderRadius="m"
+      // style={{ opacity }}
+      padding="m"
+      backgroundColor="silver"
+      gap="s"
+      flexDirection="row"
+      justifyContent="center"
+      alignContent="center"
+    >
+      <AnimatedBox borderRadius="full" backgroundColor="gray" width={70} height={70} />
+      <AnimatedBox gap="s" width="55%">
+        <AnimatedBox borderRadius="s" backgroundColor="gray" height={25} width="100%" />
+        <AnimatedBox borderRadius="s" backgroundColor="gray" height={25} width="80%" />
+      </AnimatedBox>
+      <AnimatedBox backgroundColor="gray" borderRadius="m" width={70} height={70} />
+    </AnimatedBox>
+  );
+};
+
+export default NotificationSkeleton;
