+++ frontend/app/(app)/post-creation/components/photo.tsx
@@ -1,6 +1,6 @@
 import { Box } from "@/design-system/base/box";
 import { Theme } from "@/design-system/base/theme";
-import { Icon } from "@/design-system/components/ui/icon";
+import { Icon } from "@/design-system/components/shared/icons/icon";
 import { useTheme } from "@shopify/restyle";
 import { Image, Pressable } from "react-native";
