+++ frontend/app/(auth)/check-email.tsx
@@ -5,8 +5,8 @@ import { SafeAreaView } from "react-native-safe-area-context";
 import { useState, useEffect } from "react";
 import { BaseButton } from "@/design-system/base/button";
 import { useUserState } from "@/auth/provider";
-import { TextButton } from "@/design-system/components/ui/text-button";
 import { router } from "expo-router";
+import { TextButton } from "@/design-system/components/shared/buttons/text-button";
 
 const CheckEmail = () => {
   const [isDisabled, setIsDisabled] = useState(false);
