+++ frontend/design-system/components/shared/avatar.tsx
