+++ frontend/design-system/components/shared/icons/icon.tsx
