+++ frontend/design-system/components/shared/progress-bar.tsx
